test(plugin): 合并卡片能力回归用例
